// appth/src/app/models/DinamizadoraGestacionBusqueda.ts
export class DinamizadoraGestacionBusqueda {
    id: number;
    primerNombre: string;
    segundoNombre: string;
    primerApellido: string;
    segundoApellido: string;
    tipoDocumento: string;
    nroDocumento: string;
    cargo: string;
    area: string;
    contrato: number;
    archivo: string;

    fecha: Date;

    fechaprobabaleparto: string;
    fechafum: string;
    tomavacaciones: string;
    descripcion: string;


}